Allow PoreModel to infer alphabet at read time
Adds a routine to nanopore_alphabet to loop over the two alphabets
and return the first one which contains all of given nucleotides;
change poremodel file-input constructors to cache list of kmers,
construct all seen nucleotides, and then call this to choose an
alphabet.  If an alphabet is supplied to the constructor, that is
used instead.